refactor(walletconnect): move WalletConnect code to its own assembly
WalletConnect no longer needs to be in the CareBoo.AlgoSdk assembly because of the formatter
refactors. It is now in the assembly CareBoo.AlgoSdk.WalletConnect. WalletConnect tests were also
moved to their own test assembly. It may be good to move it to its own repository in the future.

BREAKING CHANGE: Code in the AlgoSdk.WalletConnect namespace are now in the
CareBoo.AlgoSdk.WalletConnect assembly. Additionally, Hex has been moved to the AlgoSdk namespace
and remains in the CareBoo.AlgoSdk assembly.